requirements.txt
gradio
numpy
Pillow
torch
timm
transformers